refactor(components)!: 🔄 rename DropZoneInput to DropZone for better clarity feat(components): ✨ improve file upload component structure
